improved the score function
I added that the score function keeps into account the difference in score at the current state, as well as potential points earned by taking a specific move in the current turn. The tree now also updates each node's game_state.scores in add_layer_here. I also changed the maximize value of the root to True, as it seems to refer to whether the score of its children need to be maximized, rather than its own score